Change data type

+++ src/main/java/com/example/magneto/API/Models/StatsModel.java
@@ -4,17 +4,17 @@ import com.example.magneto.Model.Stats;
 
 public class StatsModel {
 
-    private int count_mutant_dna;
+    private long count_mutant_dna;
 
-    private int count_human_dna;
+    private long count_human_dna;
 
     private double ratio;
 
-    public int getCount_human_dna() {
+    public long getCount_human_dna() {
         return count_human_dna;
     }
 
-    public void setCount_human_dna(int count_human_dna) {
+    public void setCount_human_dna(long count_human_dna) {
         this.count_human_dna = count_human_dna;
     }
 
@@ -26,12 +26,12 @@ public class StatsModel {
         this.ratio = ratio;
     }
 
-    public int getCount_mutant_dna() {
+    public long getCount_mutant_dna() {
 
         return count_mutant_dna;
     }
 
-    public void setCount_mutant_dna(int count_mutant_dna) {
+    public void setCount_mutant_dna(long count_mutant_dna) {
         this.count_mutant_dna = count_mutant_dna;
     }
